avoid displaying error from git symbolic-ref -q HEAD
Usually this won't fail even if .git/HEAD is not set or points to a ref
that doesn't exist. However, early in clone, it contains
"ref: refs/heads/.invalid" which causes an error "fatal: No such ref: HEAD"

When cloning from a special remote, git-remote-annex output that once
per bundle.